Yes It's true. I really did write over 8,000 lines of code. No but the project is finished I think. Everything works as it should. Only PNG's can be uploaded but whatever live with it.

+++ src/Physics/Vector.hpp
@@ -7,7 +7,7 @@ struct Vector {
     float x;
     float y;
     Vector() = default;
-    Vector(const double x,const double y) : x(x), y(y) {}
+    Vector(const float x, const float y) : x(x), y(y) {}
 
     bool operator==(const Vector& v) const noexcept {
         return v.x == this->x && v.y == this->y;
